Move helpers_test into a package to allow re-use.
